refactor: split world and renderer, refactor the app structure
