--- src/utils/brush.rs
@@ -1,6 +1,0 @@
-use bevy_color::LinearRgba;
-
-
-pub struct Stroke {
-    
-}
