+++ src/world/canvas/shader/blend.wgsl
